+++ src/tests/LagrangeTest.java
@@ -21,17 +21,17 @@ public class LagrangeTest {
     @Test
     public void testInterpolate() {
         final RandPoly poly1 = new RandPoly(43, 2);
-        final int[] y1 = poly1.computeShares(3);
+        final long[] y1 = poly1.computeShares(3);
         final int[] x1 = new int[] {1, 2, 3};
-        final int secret = poly1.modCompute(0);
-        final int rsecret = Lagrange.interpolate(x1, y1);
+        final long secret = poly1.modCompute(0);
+        final long rsecret = Lagrange.interpolate(x1, y1);
         assertEquals("Interpolation returned an incorrect result with degree 2!", secret, rsecret);
 
         final RandPoly poly2 = new RandPoly(78, 11);
-        final int[] y2 = poly2.computeShares(12);
+        final long[] y2 = poly2.computeShares(12);
         final int[] x2 = new int[] {1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12};
-        final int secret2 = poly2.modCompute(0);
-        final int rsecret2 = Lagrange.interpolate(x2, y2);
+        final long secret2 = poly2.modCompute(0);
+        final long rsecret2 = Lagrange.interpolate(x2, y2);
         assertEquals("Interpolation returned an incorrect result with degree 11!", secret2, rsecret2);
     }
 
@@ -41,7 +41,7 @@ public class LagrangeTest {
     @Test(expected = IllegalArgumentException.class)
     public void testIllegalArgument() {
         final int[] x = new int[] {1, 2, 3};
-        final int[] y = new int[] {1, 2};
+        final long[] y = new long[] {1, 2};
         Lagrange.interpolate(x, y);
     }
 }
